requirements.txt
discord.py
openai
python-dotenv
notion-client
webcolors